fixed naming a structure of components
