+++ src/components/mattressList/index.jsx
@@ -1,156 +1,140 @@
-import React from "react";
-import { HelmetDatoCms } from "gatsby-source-datocms";
-import styled from "styled-components";
+import React, { useReducer } from "react";
 import PropTypes from "prop-types";
-import { Link } from "gatsby";
+import styled from "styled-components";
+import Header from "./CurrentSale__Header";
+import MattressThumb from "./mattThumbNail";
+import FilterSortPanel from "./FilterSortPanel";
+import filterSortReducer from "./filterSortReducer";
 import BreadCrumbs, { BreadWrapper } from "../breadCrumbs";
-import {
-  MainWrapper,
-  Wrapper,
-  // MainTitle,
-  Img,
-  StearnsImgPlaceHolder,
-  SealyImgPlace,
-} from "./mattList.styles";
-import SealyImg from "../../images/sealyLogo.png";
-import StearnsImg from "../../images/stearnsLogo.png";
-import TempurImg from "../../images/tempurLogo2.png";
-import SertaLogo from "../../images/sertaLogo.png";
 
-const NewThing = styled.header`
+const NewBread = styled(BreadWrapper)`
+  padding: 0;
+  margin-left: 0;
+  margin-right: 0;
+  @media (min-width: 568px) {
+    padding: 0;
+    margin-left: 0;
+    margin-right: 0;
+  }
+  @media (min-width: 768px) {
+    padding: 0;
+    margin-left: 0;
+    margin-right: 0;
+  }
+  @media (min-width: 1022px) {
+    padding: 0;
+    margin-left: 0;
+    margin-right: 0;
+  }
+  @media (min-width: 1300px) {
+    padding: 0;
+    margin-left: 0;
+    margin-right: 0;
+  }
+`;
+
+const MattListWrapper = styled.div`
   display: flex;
-  margin-right: 60px;
-  margin-left: 60px;
-  border-bottom: 2px solid ${props => props.theme.mainColor1};
-  margin-bottom: 20px;
-  background-color: ${props => props.theme.newColor1};
-  justify-content: space-between;
-  @media screen and (max-width: 600px) {
+  padding-left: 5px;
+  padding-right: 5px;
+  flex-direction: column;
+  .mattList__flex {
+    border-top: 8px solid ${props => props.theme.mainColor1};
+    padding-top: 20px;
+    padding-bottom: 20px;
+    display: flex;
     flex-direction: column;
-  }
-  @media screen and (max-width: 768px) {
-    padding-bottom: 10px;
-  }
-
-  a {
-    align-self: flex-end;
-    justify-self: flex-end;
-    text-decoration: none;
-    text-align: center;
-    background-color: ${props => props.theme.mainColor1};
-    border-radius: 4px;
-    border: 1px solid #ccc;
-    color: white;
-    font-family: ${props => props.theme.MainFont1};
-    text-transform: uppercase;
-    transform-style: flat;
-    transition: all ease 0.3s;
-    letter-spacing: 0.18rem;
-    font-size: 0.8rem;
-    width: 120px;
-    padding: 5px;
-    &:active {
-      box-shadow: 0 3px 0 #ccc;
-      top: 3px;
-      outline: none;
-    }
-    &:hover:enabled {
-      background-color: ${props => props.theme.mainColor2};
-      color: white;
-      cursor: pointer !important;
-    }
-    &:active:enabled {
-      background: ${props => props.theme.mainColor1} !important;
-      box-shadow: inset 0px 0px 5px #c1c1c1 !important;
-      outline: none;
-    }
-    @media (min-width: 360px) {
-      width: 130px;
-    }
-    @media (orientation: landscape) and (max-width: 568px) {
-      width: 160px;
-    }
-    @media (orientation: landscape) and (min-width: 569px) {
-      width: 180px;
-      padding: 7px;
-    }
-    @media screen and (max-width: 1024px) {
-      margin-top: 10px;
-    }
-    @media screen and (max-width: 600px) {
-      align-self: center;
-      margin-top: 0px;
+    @media screen and (min-width: 768px) {
+      flex-direction: row;
     }
   }
-  p {
-    font-size: 1.4rem;
-    line-height: 1.4rem;
-    font-family: ${props => props.theme.MainFont3};
-    font-weight: 300;
-    @media screen and (max-width: 600px) {
-      display: none;
-    }
-    @media screen and (max-width: 768px) {
-      margin-top: 0;
-      font-size: 1.2rem;
-    }
-    @media screen and (max-width: 1024px) {
-      margin-bottom: 0;
-      margin-top: 15px;
+  .mattList__grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(265px, 1fr));
+    grid-auto-rows: minmax(300px, auto);
+    grid-gap: 1rem;
+    margin-left: 7px;
+    margin-right: 7px;
+    @media screen and (min-width: 768px) {
+      margin: 0;
+      width: 80%;
     }
   }
 `;
 
-const MattressList = ({
-  children,
-  seo,
-  brandImgAlt,
-  headerText,
+const MattList = ({
+  mattresses,
+  title,
+  description,
+  breadCrumbs,
   brandName,
+  headerBG,
   landing,
 }) => {
-  const brandImgTag = name => {
-    if (name === "sealy")
-      return <SealyImgPlace src={SealyImg} alt={brandImgAlt} />;
-    if (name === "stearns")
-      return <StearnsImgPlaceHolder src={StearnsImg} alt={brandImgAlt} />;
-    if (name === "serta") return <Img src={SertaLogo} alt={brandImgAlt} />;
-    return <Img src={TempurImg} alt={brandImgAlt} />;
+  const initalState = {
+    mattresses,
+    beforeFilter: mattresses,
+    checkBoxs: [
+      { id: 0, value: "Extra Firm", firmness: 1 },
+      { id: 1, value: "Firm", firmness: 2 },
+      { id: 2, value: "Medium", firmness: 3 },
+      { id: 3, value: "Plush", firmness: 4 },
+      { id: 4, value: "Ultra Plush", firmness: 5 },
+    ],
+    firmnessNums: [],
   };
+  const [state, dispatch] = useReducer(filterSortReducer, initalState);
   return (
-    <MainWrapper>
-      <HelmetDatoCms seo={seo} />
-      <BreadWrapper Brands>
-        <BreadCrumbs next="Brands" here={brandName} />
-      </BreadWrapper>
-      <NewThing>
-        {brandImgTag(brandName)}
-        <div style={{ display: "flex", flexDirection: "column" }}>
-          <p>{headerText}</p>
-          {landing && (
-            <Link to={`/brands/${brandName}/landing`}>Learn more</Link>
-          )}
+    <MattListWrapper>
+      {breadCrumbs && (
+        <NewBread Brands>
+          <BreadCrumbs next="Brands" here={brandName} />
+        </NewBread>
+      )}
+      <Header
+        title={title}
+        description={description}
+        headerBG={headerBG}
+        landing={landing}
+        brandName={brandName}
+      />
+
+      <div className="mattList__flex">
+        <FilterSortPanel
+          dispatch={dispatch}
+          checkBoxs={state.checkBoxs}
+          length={state.mattresses.length}
+        />
+        <div className="mattList__grid">
+          {state.mattresses.map(mattress => (
+            <MattressThumb
+              key={mattress.id}
+              mattress={mattress}
+              url={`/brands/${mattress.brand.urlName}/${mattress.slug}`}
+            />
+          ))}
         </div>
-      </NewThing>
-      <Wrapper>{children}</Wrapper>
-      <BreadWrapper Brands Bottom>
-        <BreadCrumbs next="Brands" here={brandName} />
-      </BreadWrapper>
-    </MainWrapper>
+      </div>
+      {breadCrumbs && (
+        <NewBread Brands Bottom>
+          <BreadCrumbs next="Brands" here={brandName} />
+        </NewBread>
+      )}
+    </MattListWrapper>
   );
 };
 
-MattressList.defaultProps = {
-  landing: true,
+MattList.propTypes = {
+  title: PropTypes.string.isRequired,
+  description: PropTypes.string.isRequired,
+  mattresses: PropTypes.instanceOf(Object).isRequired,
+  breadCrumbs: PropTypes.bool,
+  brandName: PropTypes.string,
 };
 
-MattressList.propTypes = {
-  seo: PropTypes.instanceOf(Object).isRequired,
-  children: PropTypes.node.isRequired,
-  headerText: PropTypes.string.isRequired,
-  brandImgAlt: PropTypes.string.isRequired,
-  brandName: PropTypes.string.isRequired,
-  landing: PropTypes.bool,
+MattList.defaultProps = {
+  breadCrumbs: false,
+  brandName: "Nothing",
 };
 
-export default MattressList;
+export default MattList;
