+++ src/pages/brands/tempurpedic/index.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { graphql } from "gatsby";
 import { HelmetDatoCms } from "gatsby-source-datocms";
 import Layout from "../../../components/layout";
-import MattListCurrentSale from "../../../components/mattressList/currentSale";
+import MattListCurrentSale from "../../../components/mattressList";
 
 const Tempurpedic = ({ data }) => {
   const { datoCmsBrand, allDatoCmsMattress } = data;
